Move comparators under services
